# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Goでクリーンアーキテクチャを使用して。バックエンドAPIを作成する",
        "title": "ECサイトバックエンドAPI",
        "contact": {},
        "version": "v0.1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/v1/products/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザーを取得する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ユーザーID",
                        "name": "request",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.getUserResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "user.getUserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/user.userResponseModel"
                }
            }
        },
        "user.userResponseModel": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        }
    }
}